+++ CHANGELOG.md
@@ -8,3 +8,11 @@ Initial release
 
 - Autocomplete
 - Automatic undefined variable diagnostics
+
+## [0.1.1] - 2025-03-03
+
+### Changes
+
+- Updated LSP Backend to 0.2.0-alpha
+- Added a little download indicator when the extension is downloading the LSP, if its missing
+- Cleanup of old downloaded LSP versions